requirements.txt
# keep alphabetic
absl-py==2.1.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
lxml==5.1.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
setuptools==59.6.0
soupsieve==2.5
urllib3==2.2.1
xmltodict==0.13.0
